+++ src/dlls/hornet.h
@@ -1,9 +1,9 @@
 /***
 *
 *	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
-*	
-*	This product contains software technology licensed from Id 
-*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc. 
+*
+*	This product contains software technology licensed from Id
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc.
 *	All Rights Reserved.
 *
 *   Use, distribution, and modification of this source code and/or resulting
@@ -21,9 +21,9 @@
 //=========================================================
 #define HORNET_TYPE_RED			0
 #define HORNET_TYPE_ORANGE		1
-#define HORNET_RED_SPEED		(float)600
-#define HORNET_ORANGE_SPEED		(float)800
-#define	HORNET_BUZZ_VOLUME		(float)0.8
+#define HORNET_RED_SPEED		600.f
+#define HORNET_ORANGE_SPEED		800.f
+#define	HORNET_BUZZ_VOLUME		0.8f
 
 extern int iHornetPuff;
 
@@ -48,7 +48,7 @@ public:
 	void EXPORT TrackTouch ( CBaseEntity *pOther );
 	void EXPORT DartTouch( CBaseEntity *pOther );
 	void EXPORT DieTouch ( CBaseEntity *pOther );
-	
+
 	int TakeDamage( entvars_t *pevInflictor, entvars_t *pevAttacker, float flDamage, int bitsDamageType );
 
 	float			m_flStopAttack;
